refactor: Move IMediator interface to abstractions package
