+++ src/DispatchR/Mediator.cs
@@ -7,33 +7,6 @@ using Microsoft.Extensions.DependencyInjection;
 
 namespace DispatchR;
 
-public interface IMediator
-{
-    TResponse Send<TRequest, TResponse>(IRequest<TRequest, TResponse> request,
-        CancellationToken cancellationToken) where TRequest : class, IRequest;
-
-    IAsyncEnumerable<TResponse> CreateStream<TRequest, TResponse>(IStreamRequest<TRequest, TResponse> request,
-        CancellationToken cancellationToken) where TRequest : class, IStreamRequest;
-
-    ValueTask Publish<TNotification>(TNotification request, CancellationToken cancellationToken)
-        where TNotification : INotification;
-    
-    /// <summary>
-    /// This method is not recommended for performance-critical scenarios.  
-    /// Use it only if it is strictly necessary, as its performance is lower compared  
-    /// to similar methods in terms of both memory usage and CPU consumption.  
-    /// </summary>
-    /// <param name="request">
-    /// An object that implements INotification
-    /// </param>
-    /// <param name="cancellationToken"></param>
-    /// <returns></returns>
-    [Obsolete(message: "This method has performance issues. Use only if strictly necessary", 
-        error: false, 
-        DiagnosticId = Constants.DiagnosticPerformanceIssue)]
-    ValueTask Publish(object request, CancellationToken cancellationToken);
-}
-
 public sealed class Mediator(IServiceProvider serviceProvider) : IMediator
 {
     public TResponse Send<TRequest, TResponse>(IRequest<TRequest, TResponse> request,
@@ -73,6 +46,19 @@ public sealed class Mediator(IServiceProvider serviceProvider) : IMediator
         }
     }
 
+    /// <summary>
+    /// This method is not recommended for performance-critical scenarios.  
+    /// Use it only if it is strictly necessary, as its performance is lower compared  
+    /// to similar methods in terms of both memory usage and CPU consumption.  
+    /// </summary>
+    /// <param name="request">
+    /// An object that implements INotification
+    /// </param>
+    /// <param name="cancellationToken"></param>
+    /// <returns></returns>
+    [Obsolete(message: "This method has performance issues. Use only if strictly necessary", 
+        error: false, 
+        DiagnosticId = Constants.DiagnosticPerformanceIssue)]
     public async ValueTask Publish(object request, CancellationToken cancellationToken)
     {
         ArgumentNullException.ThrowIfNull(request);
