+++ src/DispatchR/TypeForwarders.cs
@@ -1,0 +1,3 @@
+using System.Runtime.CompilerServices;
+
+[assembly:TypeForwardedTo(typeof(DispatchR.IMediator))]
